+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 0.0.8 (wip)
 - global radar page cosmetics
+- copy to clipboard message cosmetics
   
 # Version 0.0.7
 - load max icon in fleet overview page